Changes to array_run and cluster_run

+++ array_run.py
@@ -187,8 +187,8 @@ class array_run(object):
         if self.run_in_cluster==1 and self.cluster_start_idx == -1 and self.cluster_step == -1: # this runs to run the Cxsystems over the cluster
             self.total_configs = len(self.df_anat_final_array)* self.trials_per_config
             self.config_per_node = self.total_configs / self.cluster_number_of_nodes
-            self.clipping_indices = np.arange(0, self.total_configs, self.config_per_node)[:self.total_configs / self.config_per_node]
-            cluster_run.cluster_run(self,anat_file_address,physio_file_address)
+            self.clipping_indices = np.arange(0, self.total_configs, self.config_per_node)[:int(self.total_configs / self.config_per_node)]
+            cluster_run.cluster_run(self, anat_file_address, physio_file_address)
             return
         if cluster_start_idx != -1 and cluster_step != -1: # this runs in cluster
             self.spawner(self.cluster_start_idx,self.cluster_step)
